+++ terraform/modules/aws-eks-aurora-cluster/outputs.tf
@@ -14,6 +14,10 @@ output "database_name" {
   value = var.database_name
 }
 
+output "master_username" {
+  value = var.master_username
+}
+
 output "security_group_id" {
   value = aws_security_group.this.id
 }